Remake files in tsx
